[Metrics] Hide deprecated metrics with gpu_ prefix
In #18354, these metrics were deprecated, and the change
was included in the v0.9.2 release.

We probably should only deprecate things in a v0.N.0 minor
release, so let's say these were deprecated in v0.10.0.

According to https://docs.vllm.ai/en/latest/usage/metrics.html:

> Note: when metrics are deprecated in version X.Y, they are hidden in
>  version X.Y+1 but can be re-enabled using the
>  --show-hidden-metrics-for-version=X.Y escape hatch, and are then
>  removed in version X.Y+2.

The deprecated metrics should be hidden in the v0.11.0 release,
but with a --show-hidden-metrics-for-version=0.10 escape hatch.

They should then be removed in the v0.12.0 release.

Signed-off-by: Mark McLoughlin <markmc@redhat.com>