Add structured outputs / guided decoding / logprobs support with host-side sampling (#171)
Implements a compatibility-focused sampling mode in tt_model_runner. It hooks into vllm's full logit_processor and sampler stack and enables:

structured outputs / guided decoding
logprobs (not prompt_logprobs yet, but it's a necessary step for that one)
multiple other sampling options
The limitation is that it only works with host-side sampling. LLama 70B on Galaxy always enables device sampling and is not supported in that mode.

Unless force-enabled by an override-tt-config setting, the compat sampling is enabled per-batch when sampling arguments not supported by our host sampling implementation are detected.