[Core] Get num_encoder_tokens from scheduler config
This is the same change that was made in #24866. In that PR, it was
pointed out that this code:

    MULTIMODAL_REGISTRY.get_encdec_max_encoder_len(...)

is much slower than getting the same value that's cached on the
scheduler config:

    scheduler_config.max_num_encoder_input_tokens

This PR makes the change to more spots: the scheduler, kv cache
manager, and gpu model runner.

Related to issue #24946.

Signed-off-by: Russell Bryant <rbryant@redhat.com>