fix(executor): Handle TypeError during graceful shutdown in ExecutorBase.__del__
During the Python interpreter shutdown process, the order of object garbage collection is not guaranteed. This can lead to a race condition where the `shutdown()` method is called in `ExecutorBase.__del__`, but some dependent objects (like `collective_rpc` in `UniprocExecutor`) have already been destroyed and set to `None`.
This results in a `TypeError: 'NoneType' object is not callable`, printing an ignored exception traceback to the console, which can be confusing.
This commit wraps the `self.shutdown()` call within a `try...except` block to catch `TypeError`. This ensures a clean and silent exit without raising an exception if the shutdown process is interrupted by garbage collection.

Platform Context:
- Arch: sg2044riscv64
- OS: EulixOS 3.0
- GCC: 15.1
- PyTorch: 2.8.0
- Python: 3.11.6

    Signed-off-by: lyd1992 <liuyudong@iscas.ac.cn>
    Signed-off-by: ihb2032 <1355790728@qq.com>