[Bugfix] [Frontend] Cleanup gpt-oss non-streaming chat tool calls
This fixes an issue identified in #22337 that was not entirely fixed
by #22386 around tool call parsing in gpt-oss models in the
non-streaming case and the model output returned to the user.

The main change here is to remove the parsed tool call out of the
ChatCompletion message `content`, so that the generated tool call only
shows up in its parsed form as an entry in `tool_calls` instead of in
both `tool_calls` and `content`.

A small related change is to ensure we're not sending newline
characters in the JSON arguments string of the parsed tool call, since
we don't do this for other tool calls.

Together these should get non-streaming tool calls via the Chat
Completions API working for gpt-oss models for typical
use-cases. There may still be some edge cases the
openai_tool_parser.py and/or serving_chat.py code need to handle, but
this at least closes some known gaps.

A couple of unit tests were tweaked here to test for both of these
fixes. I ensured the tests failed before my fixes, and that they now
pass with the fixes.

Signed-off-by: Ben Browning <bbrownin@redhat.com>