[BugFix] Use async Mistral Tokenizer in Chat Completions
When using `--tokenizer-mode mistral` with Chat Completions, large
inputs could block the event loop, causing delays in processing of other
requests. This adjusts the usage of that tokenizer during chat request
preprocessing to use pre-existing but unused `_tokenizer_executor` in
OpenAIServing to run these blocking operations in a background thread as
opposed to running them directly in the critical path of the server
event loop.

Signed-off-by: Ben Browning <bbrownin@redhat.com>