Add model validation plugin and registry and validate local models
Add a model validation plugin registry where classes implementing the
ModelValidationPlugin interface can be registered.

Enable the validating on local models that have already been downloaded
by the user.

Add a test case with an already downloaded model whose config.json is
unmodified so that a ModelConfig can be created from it.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>