Enable model validation on downloaded models
Implement a method 'validate' in the BaseModelLoader that first checks
whether any plugin requests to validate the given model and then possibly
downloads all the model files, including the signature. For this, query
the subclass of BaseModelLoader for its download type. Support validation
of local models and those downloaded from Huggingface Hub.

Add a test case.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>