Refactor to use make_async utils method
After digging a bit more, I discovered we have an existing `make_async`
method that encapsulates turning a block call into an async one. So, I'm
simplifying this change to take advantage of that new method. In doing
so, I needed to pull the if statement that determines which chat
template to apply out of `_preprocess_chat` and into its own
`_apply_chat_template` to make testing of this newly simplified path
easier without having to do a lot of mocking for all the other concerns
covered in the larger `_preprocess_chat` method.

Signed-off-by: Ben Browning <bbrownin@redhat.com>