[Frontend] Respect Chat Completion parallel_tool_calls param
Our Chat Completions API exposed a `parallel_tool_calls` parameter but
didn't actually use that to control the number of tool calls returned.
This fixes that, so that if a user sets `parallel_tool_calls` to `False`
then we'll only return a single tool call for streaming or non-streaming
responses instead of allowing more than one.

The default for this parameter in the Chat Completions API spec is
`True`, and the previous behavior of vLLM acted like this was set to
`True`. So, even though in this change the default in the code changes
from `False` to `True`, from a user's point-of-view and
backwards-compatibility perspective the default has not changed. We were
ignoring the value entirely before and acting like it was always set to
`True`.

This does not attempt to stop generation once the first tool call is
generated. Instead, it simply drops any tool calls after the first when
sending the response back to the client if `parallel_tool_calls` is
`False`. Figuring out how or if we could stop the generation early to
avoid wasting GPU cycles is above my current knowledge of the internals
here.

Signed-off-by: Ben Browning <bbrownin@redhat.com>