[CI/Build] Add common tool call parser test suite
This adds a common test suite for tool call parsers and wires all of the
existing tool call parsers that had no tests into the common suite. It
doesn't yet adapt existing tool call parser tests to fit into the common
suite nor augment tool call parsers that already had tests with the new
set of common tests. Those tasks can come later, as this PR is already
quite large.

Not all of the existing tool call parsers can pass every test in the
common test suite. The ones that are not passing today are marked as
xfail, and those represent opportunities to identify and fix gaps in all
of these tool call parsers in the future until we get down to zero
expected to fail tests within the common suite for each parser.

Given how many tests are here, the default_tokenizer fixture used by
tests when tokenizing strings was also adjusted to be module-scoped, so
we don't create a new version of that for every single test.

Signed-off-by: Ben Browning <bbrownin@redhat.com>