CLI: Ignore SIGTSTP signal (Ctrl-Z)
SIGTSTP is a user-interaction signal meant to notify the program that it will be suspended. It is not appropriate to exit the program when receiving a SIGTSTP. It would be appropriate to release a lock, set state so that future actions can occur upon process resumption, etc. but the current behavior is probably not desireable.

Signed-off-by: Thom Mahoney <thommahoney@google.com>