[Bugfix] Catch and log invalid token ids in detokenizer
This is an update to the "workaround" added in https://github.com/vllm-project/vllm/pull/24351.

That PR insulates against occasional negative token ids that can be produced occasionally, though we still don't know the root cause.

With the update to tokenizers 0.22.1, this error manifests as a TypeError rather than an OverflowError, so the patch needs to be updated to account for this.

Signed-off-by: Nick Hill <nhill@redhat.com>