[Chore]: Stream tokens vs characters in tool call parser tests
In the `run_tool_extraction_streaming` test function that helps us test
streaming tool call scenarios, we were streaming individual characters
at a time into the extract_tool_calls_streaming methods of tool call
parsers when the `model_deltas` passed in was just a string. This is
incorrect, as we should be streaming at least 1 token in with every
request instead of individual characters.

This is laying the groundwork for a future PR where I'm adding tests to
all the tool call parsers that do not have any tests. But, since this
requires tweaking a couple of the existing parser tests to pass in real
tokenizers instead of a simple mock, I broke this out into its own PR so
that the one that adds more tool call tests can be a purely additive
change.

This is marked as a chore because it's just making our streaming tests
more true to what really happens, and is required before I can add
working tests for some of the untested parsers today.

Signed-off-by: Ben Browning <bbrownin@redhat.com>