Add tests for chunked prefill and prefix cache with causal pooling models
This test uses Qwen3-Embedding-0.6B which has causal attention
and LAST pooling, and therefore supports these features.
To make these verifications, the test creates an interceptor
Pooler which verifies if the prompt processing is done in
one go or piecewise. For prefix caching it verifies if
the number of tokens that it has seen is less than expected.

Signed-off-by: Max de Bayser <mbayser@br.ibm.com>
Co-authored-by: Ayush Singh <ayush1009208@gmail.com>