Add support for the /rerank endpoint in vllm bench serve
The /rerank API can be support both by embedding models
and native reranker models. However, with reranker
models the query is concatenated with each document with
a separator token in between. Therefore the amount
of tokens that passes through the model has to be accounted
for differently in each case.
Because of these details, this PR, adds a specialized random
dataset to generates requests which send the expected amount
of tokens. So when the use sets `random-input-len`, `num-prompts`
and `random-batch-size`, in both cases we will generate requests
such that the total amount of tokens is prompts*input-len in
batches of size batch-size*input-len.

Signed-off-by: Max de Bayser <mbayser@br.ibm.com>