Notice for deprecation
Having a deprecated tool front and center in the quantization section of the vllm docs may be contributing to the number of users we see struggling to get AWQ working in vllm (https://discuss.vllm.ai/t/a-bit-of-frustration-with-quantization/1720). Directing them to maintained frameworks first has the potential to reduce this.

Signed-off-by: HDCharles <39544797+HDCharles@users.noreply.github.com>