fix: a more robust way to ensure only a single video item is processed at a time
Instead of relying on input shapes and that the first dim is the batch
dim, just process a single item. Changes in
group_mm_kwargs_by_modality() which now has merge_by_field_config=True
for almost all relevant models mean that the batch dimension doesn't
exist and frames from different videos are gathered in the same dim.
That's why we need a more robust way to deal with processing just a
single video

Signed-off-by: Tomer Asida <57313761+tomeras91@users.noreply.github.com>