GPUModelRunner: Support cross-layer KV blocks
This commit enables the GPU model runner to allocate the KV cache tensors,
so that the KV data for all layers will be contiguous per block.
This can yield a significant speed up the transfer time of KV transfers (e.g. X4),
such in the case of using NixlConnector or OffloadingConnector.
Currently, this new layout is disabled by default, and will only be enabled when using
a connector which explicitly prefers this new layout.
Also, this new layout is currently only supported for uniform (non HMA) models.

Signed-off-by: Or Ozeri <oro@il.ibm.com>