honor --mm_encoder_attn_backend when used (#27124)
Summary:
Pull Request resolved: https://github.com/vllm-project/vllm/pull/27124

In https://github.com/vllm-project/vllm/pull/26104, some changes were made in layer.py that resulted in always trying to switch to FA backend for ViT, even when  `VLLM_ATTENTION_BACKEND` is set.

This broke Meta's internal AMD pipelines as it is not desired nor expected behavior. With this change, the models that were changed in the offending PR can explicitly opt-in to this behavior.

Reviewed By: Prowindy

Differential Revision: D84946967