[Frontend] Make RequestIdMiddleware return the internal request_id
RequestIdMiddleware can be enabled using --enable-request-id-headers
and it will add an X-Request-ID header to the response. Currently:

- If an `X-Request-ID` header is provided in the client request,
  we prepend a prefix ('cmpl-' for example) but we do not include
  this in the response header.
- If the client does not provide a request ID, we include a random
  ID in the response that has no relation to the request ID internally
  and therefore is useless for correlating logs.

In #8672 we added RequestResponseMetadata for external middlewares and
we do not use this internally currently. We can rely on it to allow
us to return the actual internal request ID in the response header.

Example:

```
$ curl -v http://127.0.0.1:8000/v1/completions ...
...
< x-request-id: cmpl-f7fee54511494349b0f0e5c6a079573a
<
{"id":"cmpl-f7fee54511494349b0f0e5c6a079573a", ...}
```

Signed-off-by: Mark McLoughlin <markmc@redhat.com>