Perf gain because of block alignment
Summary:
Following patch is from Aliasger Zaidy(azaid) and Shucai Xiao(scxiao) from AMD, and overall efforts for the integration is guided by Xiaozhu Meng(mxz297) from Meta, it boosts the performance for fused_moe kernel.
We pad to 128 for MI300 to avoid masked loads.
We pad to 256 for MI355 because we use scale preshuffling on 355 and padding to 256 is needed to enable correct preshuffle arrangement

10% Performance boost is achieved for gptoss120b on AMD mi300 machine.

Test Plan:
No eval regression is observed.

| Effort Level |  Score | Characters | Chars Std | Score Std |
|--------------|-------|------------|-----------|-----------|
| Low  | 0.51 | 1577.26 | 1001.32 | 0.49 |
| Medium  | 0.79 | 1991.975 | 785.68 | 0.40 |
| High  | 0.916 | 2568 | 1029.9 | 0.28 |

| Effort Level |  Score | Characters | Chars Std | Score Std |
|--------------|-------|------------|-----------|-----------|
| Low  | 0.51 | 1570.26 | 1001.32 | 0.49 |
| Medium  | 0.79 | 1990.975 | 780.68 | 0.40 |
| High  | 0.916 | 2508 | 1020.9 | 0.28 |

Signed-off-by: Smit Kadvani <smit.kadvani@gmail.com>