Enable device sampling with per-batch host sampling fallback when required (#232)
Extend vLLM v0 by adding automatic compat sampling fallbacks to device sampling. This offers substantial performance benefits when not all requests require advanced features like structured outputs.
always_compat_sampling should now never be set in production - whenever something otherwise unsupported, such as

structured outputs
non-greedy sampling for models that don't support it on device
non-uniform top-k top-p on host or with a model that doesn't support it
is detected, the system will switch to host compat sampling.
This also means that we never override the temperature, p or k to force it to be uniform across batch.

Also contains a fix for #229 (it will fallback to compat sampling if non-uniform is not supported on device or device sampling is disabled).