[Bugfix] Parse gpt-oss refusals w/ newer openai-harmony
The output generated by gpt-oss models does not always strictly follow
its expected harmony chat template format. This commonly - but not
exclusively - happens when gpt-oss-120b generates refusals for content
that violates its built-in safety guidelines.

To fix this, a non-strict mode was added to the openai-harmony library
to allow attempted recovery of malformed message headers in the model
output, such as a missing `<|message|>` special token before the
assistant text.

This will resolve some cases where the error
`openai_harmony.HarmonyError: unexpected tokens remaining in message
header` was previously thrown. It will not resolve all of those, as not
every malformed message output can be recovered. Other ongoing work
around using structured output for the Harmony format can help prevent
these kinds of things in the first place, once that work lands and in
the cases where the user and/or server decide to enable it.

I believe it should be safe to enable this non-strict mode by default in
vLLM, as the code paths that enables in the openai-harmony library only
gets triggered once it's already detected malformed output. So, there
shouldn't be any performance penalty in the common case. And, in the
event that the malformed content cannot be properly recovered, the
openai-harmony library will still end up throwing an error.

This is related to vllm-project/vllm#23567 as well as openai/harmony#80.

Signed-off-by: Ben Browning <bbrownin@redhat.com>