kv_offloading: Fix bug in loading of partial cpu blocks
This commit fixes a bug when trying to load from the middle of a CPU block.
This can happen if cpu_block_size > gpu_block_size,
and there's both a cpu and gpu (prefix cache) hit,
where the gpu hit ends in the middle of a cpu block.
Before this commit, the code tried to wrongfully address the other direction,
storing to the middle of a cpu block. But this is impossible since the offloading connector
always stores full CPU blocks.

Signed-off-by: Or Ozeri <oro@il.ibm.com>