block_pool: sort the list of allocated free blocks
Sorting the blocks returned from the free list, increases the chance
that blocks that correspond to the same request and therefore KV cache
content, will also be sequential in the device memory. This helps later
for offload engines that want to write out that KV cache to file with a
smaller vector of scatter gather, which can increase IO size and
therefore offload throughput.

Signed-off-by: Dan Aloni <dan.aloni@vastdata.com>