[NIXL] Add remote_request_id to kv_transfer_params
Include the internal request ID that the prefill instance is
expecting the decode instance to send it in the NIXL notification.

Right now, we rely on the proxy supplying the ID via X-Request-ID
and that prefill and decode will mangle this ID in identical ways.
This is obviously quite brittle, and P should be explicit about what
ID it expects from D.

Relates to #27987 - adding a random prefix to client-provided
request IDs.

Signed-off-by: Mark McLoughlin <markmc@redhat.com>